requirements.txt
aiohttp>=3.9
ccxt>=4.0
psycopg2-binary>=2.9
pydantic-settings>=2.0
python-dotenv>=1.0
websockets>=12.0
requests>=2.31
